Oh you can use flippin cactus for rails now.
...and I think I can do more than this.
TODO : Do not update observer with flippincactus
Trapdoors, Repeaters, Hoppers (only side), Fence gates has potential to be fixed too. 
Currently its only for rails, so it only flips once so for loop is not really needed.
It cant get server's flippincactus information, so you need to manually get cactus on hand and turn the option on.